2059: Use repositories
